backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
backports.asyncio.runner==1.2.0
certifi==2026.4.22
click==8.3.3
colorama==0.4.6
deepdiff==9.0.0
exceptiongroup==1.3.1
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
Mako==1.3.12
MarkupSafe==3.0.3
orderly-set==5.5.0
packaging==26.2
pluggy==1.6.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
